+++ src/App.js
@@ -2,14 +2,14 @@ import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { Routes } from './Routes';
+import { Routes } from './routes';
 import './App.scss';
-import AppTabs from './SmartComponents/AppTabs/AppTabs';
+import AppTabs from './smart-components/app-tabs/app-tabs';
 import { Main, PageHeader } from '@redhat-cloud-services/frontend-components';
 import { Title } from '@patternfly/react-core';
 import { NotificationsPortal } from '@redhat-cloud-services/frontend-components-notifications/';
 import '@redhat-cloud-services/frontend-components-notifications/index.css';
-import { AppPlaceholder } from './PresentationalComponents/Shared/LoaderPlaceholders';
+import { AppPlaceholder } from './presentational-components/shared/loader-place-holders';
 
 class App extends Component {
   state = {
