--- src/PresentationalComponents/Shared/404Route.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-/**
- * Just a placeholder component. Will w8 for some designs.
- */
-const NoMatch = () => (
-  <div>
-    <h1>
-      Page you are looking for does not exist.
-    </h1>
-  </div>
-);
-
-export default NoMatch;
