--- src/PresentationalComponents/User/UsersFilterToolbar.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import FilterToolbarItem from '../Shared/FilterToolbarItem';
-
-const UsersFilterToolbar = ({ onFilterChange, filterValue, ...props }) => (
-  <FilterToolbarItem { ...props } searchValue={ filterValue } onFilterChange={ onFilterChange } placeholder={ 'Find a User' }/>
-);
-
-UsersFilterToolbar.propTypes = {
-  onFilterChange: PropTypes.func.isRequired,
-  filterValue: PropTypes.string
-};
-
-export default UsersFilterToolbar;
