--- src/Utilities/Constants.js
@@ -1,2 +1,0 @@
-export const RBAC_API_BASE = `${process.env.BASE_PATH}/rbac/v1`;
-
