--- src/Utilities/store.js
@@ -1,22 +1,0 @@
-
-import promiseMiddleware from 'redux-promise-middleware';
-import ReducerRegistry, { applyReducerHash } from '@redhat-cloud-services/frontend-components-utilities/files/ReducerRegistry';
-import { notifications, notificationsMiddleware } from '@redhat-cloud-services/frontend-components-notifications/';
-
-import reduxLogger from 'redux-logger';
-import thunk from 'redux-thunk';
-import userReducer, { usersInitialState } from '../redux/reducers/userReducer';
-import groupReducer, { groupsInitialState } from '../redux/reducers/groupReducer';
-
-const registry = new ReducerRegistry({}, [ thunk, promiseMiddleware(), notificationsMiddleware({
-  errorTitleKey: [ 'message' ],
-  errorDescriptionKey: [ 'errors', 'stack' ]
-}), reduxLogger ]);
-
-registry.register({
-  userReducer: applyReducerHash(userReducer, usersInitialState),
-  groupReducer: applyReducerHash(groupReducer, groupsInitialState),
-  notifications
-});
-
-export default registry.getStore();
