--- src/redux/ActionTypes.js
@@ -1,8 +1,0 @@
-export const FETCH_USERS = 'FETCH_USERS';
-
-export const FETCH_GROUP = 'FETCH_GROUP';
-export const FETCH_GROUPS = 'FETCH_GROUPS';
-
-export const ADD_GROUP = 'ADD_GROUP';
-export const UPDATE_GROUP = 'UPDATE_GROUP';
-export const REMOVE_GROUP = 'REMOVE_GROUP';
